requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-import-export==2.9.0
et-xmlfile==1.1.0
executing==1.1.0
Faker==15.0.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
MarkupPy==1.14
matplotlib-inline==0.1.6
odfpy==1.4.1
openpyxl==3.0.10
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tablib==3.2.1
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
